requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
mangum==0.17.0